Parameters of build.py are passed by command argument. proj.tmpl will be deprecated.
